Replace #[mantle::service] with compiler plugin (#102)
* Start moving macro to compiler plugin

* s/idl-gen/mantle-build/g

* Impl rpc call tree macro in plugin

* Ctor payload

* Shuffle files

* Remove U256 type

* Shuffle more files

* u32 return types in blockchain traits

* Add functionality to Address

* mantle ffi

* Update memchain tests to use new Address

* Update bcfs tests to use new Address

* Events

* Iron out bugs in mantle-build

* Old school proc macro style

* Namespace mantle externs

* Use multi-spans

* Update idl_gen test

* Remove xcc tests which no longer work but will soon

* Bump versions

* Update compiletests

* Add test for idl-gen failure

* Update circleci

* Fix service derive for tuple structs

* Fix bcfs

* Update ballot example

* Update messaging example

* Fix bcfs

* Fix deps

* Fix service output dir

* Fix feature

* Hopefully fix tests

* Find serde

* Use tempdir

* Fix flaky run-pass compiletest

* Update mantle-build readme

* Apply suggestions from code review

* Update git deps

+++ memchain/src/lib.rs
@@ -99,7 +99,7 @@ impl<'bc> Blockchain for Memchain<'bc> {
         self.last_block().fetch_input()
     }
 
-    fn input_len(&self) -> u64 {
+    fn input_len(&self) -> u32 {
         self.last_block().input_len()
     }
 
@@ -115,7 +115,7 @@ impl<'bc> Blockchain for Memchain<'bc> {
         self.last_block().fetch_ret()
     }
 
-    fn ret_len(&self) -> u64 {
+    fn ret_len(&self) -> u32 {
         self.last_block().ret_len()
     }
 
@@ -123,7 +123,7 @@ impl<'bc> Blockchain for Memchain<'bc> {
         self.last_block().fetch_err()
     }
 
-    fn err_len(&self) -> u64 {
+    fn err_len(&self) -> u32 {
         self.last_block().err_len()
     }
 
@@ -135,7 +135,7 @@ impl<'bc> Blockchain for Memchain<'bc> {
         self.last_block().code_at(addr)
     }
 
-    fn code_len(&self, addr: &Address) -> u64 {
+    fn code_len(&self, addr: &Address) -> u32 {
         self.last_block().code_len(addr)
     }
